added full crud for books controller

+++ app/controllers/books_controller.rb
@@ -2,8 +2,18 @@ class BooksController < ApplicationController
     # index route
     get '/books' do
         books = Book.all
-        books.to_json
+        books.to_json(include: [reader: {only: [:id, name]}])
     end
+#ev
+    get '/books/:id' do
+        book = Book.find_by(params["id"])
+    #     if reader 
+            book.to_json(include: [:reader])
+    #     else 
+    #         "404 - Reader not found"
+    #     end 
+    end
+
 
         #post 'owners/:owner_id/books' do
             #reader = Reader.find_by(id: params[:reader_id])
@@ -11,20 +21,52 @@ class BooksController < ApplicationController
             #dog.to_json
         #end
 #
-    post '/books' do
-        book = Book.create(params)
-        dog.to_json
+    # post '/books' do
+    #     book = Book.create(params)
+    #     book.to_json
+    # end 
+#ev
+    post "/books" do
+        book = Book.new(params)
+        if book.save
+            book.to_json(include: [:reader])
+        else
+            {errors: book.errors.full_messages}.to_json
+        end 
     end 
 #
-    delete '/books/:id' do 
-        book = Book.find_by(id: params[:id])
-        book.destroy
-    end 
+    # delete '/books/:id' do 
+    #     book = Book.find_by(id: params[:id])
+    #     book.destroy
+    # end 
 #
-    patch 'books/:id' do
-        book = Book.find_by(id: params[:id])
-        book.update(title: params[:title])
-        book.to_json
+    # patch 'books/:id' do
+    #     book = Book.find_by(id: params[:id])
+    #     book.update(title: params[:title])
+    #     book.to_json
+    # end
+#ev
+    patch "/books/:id" do
+        
+        book = Book.find_by_id(params["id"])
+        if book.update(params)
+            book.to_json(include: [:reader])
+        else 
+            {errors: book.errors.full_messages}.to_json
+        end
     end
+
+    delete "/books/:id" do
+        book = Book.find_by_id(params["id"])
+        if book
+            book.destroy
+            book.to_json
+        else 
+            {errors: ["Book Doesn't Exist"]}.to_json
+        end 
+
+    end
+
+
 
 end
